iOS 9 iPad is good with FrontFlash + silent iOS 10 support.

+++ Tweak.h
@@ -27,7 +27,7 @@ CGFloat hue;
 CGFloat sat;
 CGFloat bri;
 
-int colorProfile;
+NSInteger colorProfile;
 
 UIColor *frontFlashColor()
 {
